firmware/raspberry-pi/requirements.txt
bleak>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
asyncio>=3.4.3
